+++ Client/Frontend/Login Management/LoginListViewController.swift
@@ -18,9 +18,9 @@ private extension UITableView {
     }
 }
 
-private let CellReuseIdentifier = "cell-reuse-id"
-private let SectionHeaderId = "section-header-id"
-private let LoginsSettingsSection = 0
+let CellReuseIdentifier = "cell-reuse-id"
+let SectionHeaderId = "section-header-id"
+let LoginsSettingsSection = 0
 
 class LoginListViewController: SensitiveViewController {
 
@@ -241,15 +241,13 @@ extension LoginListViewController: UISearchResultsUpdating {
     }
 }
 
-fileprivate var isDuringSearchControllerDismiss = false
-
 extension LoginListViewController: UISearchControllerDelegate {
     func willDismissSearchController(_ searchController: UISearchController) {
-        isDuringSearchControllerDismiss = true
+        viewModel.isDuringSearchControllerDismiss = true
     }
 
     func didDismissSearchController(_ searchController: UISearchController) {
-        isDuringSearchControllerDismiss = false
+        viewModel.isDuringSearchControllerDismiss = false
     }
 }
 
@@ -428,73 +426,5 @@ extension LoginListViewController: SearchInputViewDelegate {
     @objc func searchInputViewFinishedEditing(_ searchView: SearchInputView) {
         setupDefaultNavButtons()
         loadLogins()
-    }
-}
-
-/// Data source for handling LoginData objects from a Cursor
-class LoginDataSource: NSObject, UITableViewDataSource {
-    fileprivate let emptyStateView = NoLoginsView()
-    fileprivate var viewModel: LoginListViewModel
-
-    let boolSettings: (BoolSetting, BoolSetting)
-
-    init(viewModel: LoginListViewModel) {
-        self.viewModel = viewModel
-        boolSettings = (
-            BoolSetting(prefs: viewModel.profile.prefs, prefKey: PrefsKeys.LoginsSaveEnabled, defaultValue: true, attributedTitleText: NSAttributedString(string: Strings.SettingToSaveLogins)),
-            BoolSetting(prefs: viewModel.profile.prefs, prefKey: PrefsKeys.LoginsShowShortcutMenuItem, defaultValue: true, attributedTitleText: NSAttributedString(string: Strings.SettingToShowLoginsInAppMenu)))
-        super.init()
-    }
-
-    @objc func numberOfSections(in tableView: UITableView) -> Int {
-        if viewModel.dataSourceViewModel.loginRecordSections.isEmpty {
-            tableView.backgroundView = emptyStateView
-            return 1
-        }
-
-        tableView.backgroundView = nil
-        // Add one section for the settings section.
-        return viewModel.dataSourceViewModel.loginRecordSections.count + 1
-    }
-
-    @objc func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        if section == LoginsSettingsSection {
-            return 2
-        }
-        return viewModel.dataSourceViewModel.loginsForSection(section)?.count ?? 0
-    }
-
-    @objc func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = ThemedTableViewCell(style: .subtitle, reuseIdentifier: CellReuseIdentifier)
-
-        if indexPath.section == LoginsSettingsSection {
-            let hideSettings = viewModel.dataSourceViewModel.searchController?.isActive ?? false || tableView.isEditing
-            let setting = indexPath.row == 0 ? boolSettings.0 : boolSettings.1
-            setting.onConfigureCell(cell)
-            if hideSettings {
-                cell.isHidden = true
-            }
-
-            // Fade in the cell while dismissing the search or the cell showing suddenly looks janky
-            if isDuringSearchControllerDismiss {
-                cell.isHidden = false
-                cell.contentView.alpha = 0
-                cell.accessoryView?.alpha = 0
-                UIView.animate(withDuration: 0.6) {
-                    cell.contentView.alpha = 1
-                    cell.accessoryView?.alpha = 1
-                }
-            }
-        } else {
-            guard let login = viewModel.dataSourceViewModel.loginAtIndexPath(indexPath) else { return cell }
-            cell.textLabel?.text = login.hostname
-            cell.detailTextColor = UIColor.theme.tableView.rowDetailText
-            cell.detailTextLabel?.text = login.username
-            cell.accessoryType = .disclosureIndicator
-        }
-        // Need to override the default background multi-select color to support theming
-        cell.multipleSelectionBackgroundView = UIView()
-        cell.applyTheme()
-        return cell
     }
 }
