+++ Client/Frontend/Login Management/NoLoginsView.swift
@@ -1,10 +1,6 @@
-//
-//  NoLoginsView.swift
-//  Client
-//
-//  Created by Vanna Phong on 6/11/20.
-//  Copyright © 2020 Mozilla. All rights reserved.
-//
+/* This Source Code Form is subject to the terms of the Mozilla Public
+* License, v. 2.0. If a copy of the MPL was not distributed with this
+* file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import Foundation
